+++ samlang-core-checker/__tests__/index.test.ts
@@ -1,5 +1,5 @@
 import { Range, ModuleReference, functionType, intType } from 'samlang-core-ast/common-nodes';
-import { SourceExpressionInt } from 'samlang-core-ast/samlang-expressions';
+import { SourceExpressionInt } from 'samlang-core-ast/samlang-nodes';
 import { createGlobalErrorCollector } from 'samlang-core-errors';
 import { parseSamlangModuleFromText } from 'samlang-core-parser';
 import { mapOf, hashMapOf } from 'samlang-core-utils';
