+++ samlang-core-checker/__tests__/module-references-collector.test.ts
@@ -28,7 +28,7 @@ import {
   SourceExpressionMatch,
   SourceExpressionLambda,
   SourceExpressionStatementBlock,
-} from 'samlang-core-ast/samlang-expressions';
+} from 'samlang-core-ast/samlang-nodes';
 import { hashSetOf } from 'samlang-core-utils';
 
 import { collectModuleReferenceFromExpression } from '../module-references-collector';
