+++ samlang-core-checker/__tests__/statement-type-checker.test.ts
@@ -17,7 +17,7 @@ import {
   SourceExpressionInt,
   SourceExpressionVariable,
   SourceExpressionStatementBlock,
-} from 'samlang-core-ast/samlang-expressions';
+} from 'samlang-core-ast/samlang-nodes';
 import type { Pattern } from 'samlang-core-ast/samlang-pattern';
 import { createGlobalErrorCollector } from 'samlang-core-errors';
 import { hashMapOf, LocalStackedContext } from 'samlang-core-utils';
