+++ samlang-core-checker/expression-type-fixer.ts
@@ -10,7 +10,7 @@ import {
   isTheSameType,
   prettyPrintType,
 } from 'samlang-core-ast/common-nodes';
-import type { SamlangExpression } from 'samlang-core-ast/samlang-expressions';
+import type { SamlangExpression } from 'samlang-core-ast/samlang-nodes';
 import { assert, zip } from 'samlang-core-utils';
 
 import type { ReadOnlyTypeResolution } from './type-resolution';
