+++ samlang-core-checker/statement-type-checker.ts
@@ -4,7 +4,7 @@ import {
   SamlangValStatement,
   StatementBlock,
   SourceExpressionVariable,
-} from 'samlang-core-ast/samlang-expressions';
+} from 'samlang-core-ast/samlang-nodes';
 import type { Pattern, ObjectPatternDestucturedName } from 'samlang-core-ast/samlang-pattern';
 import type { FieldType } from 'samlang-core-ast/samlang-toplevel';
 import type { ModuleErrorCollector } from 'samlang-core-errors';
