+++ samlang-core-interpreter/__tests__/source-level-interpreter.test.ts
@@ -5,7 +5,7 @@ import {
   SourceExpressionTrue,
   SourceExpressionMethodAccess,
   SourceExpressionMatch,
-} from 'samlang-core-ast/samlang-expressions';
+} from 'samlang-core-ast/samlang-nodes';
 import { DEFAULT_BUILTIN_TYPING_CONTEXT } from 'samlang-core-checker';
 import { createGlobalErrorCollector } from 'samlang-core-errors';
 import { parseSamlangExpressionFromText, parseSamlangModuleFromText } from 'samlang-core-parser';
