+++ samlang-core-parser/__tests__/index.test.ts
@@ -1,5 +1,5 @@
 import { ModuleReference } from 'samlang-core-ast/common-nodes';
-import type { SamlangExpression } from 'samlang-core-ast/samlang-expressions';
+import type { SamlangExpression } from 'samlang-core-ast/samlang-nodes';
 import { createGlobalErrorCollector } from 'samlang-core-errors';
 
 import { parseSamlangModuleFromText, parseSamlangExpressionFromText } from '..';
