+++ samlang-core-parser/index.ts
@@ -1,5 +1,5 @@
 import type { ModuleReference } from 'samlang-core-ast/common-nodes';
-import type { SamlangExpression } from 'samlang-core-ast/samlang-expressions';
+import type { SamlangExpression } from 'samlang-core-ast/samlang-nodes';
 import type { SamlangModule } from 'samlang-core-ast/samlang-toplevel';
 import type { ModuleErrorCollector } from 'samlang-core-errors';
