+++ samlang-core-parser/samlang-parser.ts
@@ -51,7 +51,7 @@ import {
   SourceExpressionLambda,
   SourceExpressionMatch,
   SourceExpressionStatementBlock,
-} from 'samlang-core-ast/samlang-expressions';
+} from 'samlang-core-ast/samlang-nodes';
 import type { Pattern } from 'samlang-core-ast/samlang-pattern';
 import type {
   ClassDefinition,
