+++ samlang-core-printer/__tests__/printer-source-level.test.ts
@@ -2,7 +2,7 @@ import { intType, ModuleReference } from 'samlang-core-ast/common-nodes';
 import {
   SourceExpressionMethodAccess,
   SourceExpressionVariable,
-} from 'samlang-core-ast/samlang-expressions';
+} from 'samlang-core-ast/samlang-nodes';
 import { createGlobalErrorCollector } from 'samlang-core-errors';
 import { parseSamlangExpressionFromText, parseSamlangModuleFromText } from 'samlang-core-parser';
 import { checkNotNull } from 'samlang-core-utils';
