+++ samlang-core-services/language-service.ts
@@ -9,7 +9,7 @@ import {
   TypedComment,
   Sources,
 } from 'samlang-core-ast/common-nodes';
-import type { SamlangExpression } from 'samlang-core-ast/samlang-expressions';
+import type { SamlangExpression } from 'samlang-core-ast/samlang-nodes';
 import type {
   ClassDefinition,
   ClassMemberDefinition,
