+++ samlang-core-services/variable-definition-service.ts
@@ -1,5 +1,5 @@
 import type { ModuleReference, Range, Sources } from 'samlang-core-ast/common-nodes';
-import { SourceExpressionVariable, SamlangExpression } from 'samlang-core-ast/samlang-expressions';
+import { SourceExpressionVariable, SamlangExpression } from 'samlang-core-ast/samlang-nodes';
 import type { Pattern } from 'samlang-core-ast/samlang-pattern';
 import type { SamlangModule } from 'samlang-core-ast/samlang-toplevel';
 import { HashMap, LocalStackedContext, assert, error, hashMapOf } from 'samlang-core-utils';
